Refactored to use middleware
