+++ src/authentication/password.rs
@@ -1,0 +1,131 @@
+use anyhow::Context;
+use argon2::{password_hash::SaltString, Argon2, PasswordHash, PasswordHasher, PasswordVerifier};
+
+use rand::rngs::OsRng;
+use secrecy::{ExposeSecret, Secret};
+use sqlx::SqlitePool;
+
+use crate::telemetry::spawn_block_with_tracing;
+
+pub struct Credentials {
+    pub username: String,
+    pub password: Secret<String>,
+}
+
+#[derive(thiserror::Error, Debug)]
+pub enum AuthError {
+    #[error("Invalid Credentials")]
+    InvalidCredentials(#[source] anyhow::Error),
+    #[error(transparent)]
+    UnexpectedError(#[from] anyhow::Error),
+}
+
+#[tracing::instrument {
+    name = "Verify password"
+    skip(expected_password_hash, password_candidate)
+}]
+fn verify_password_hash(
+    expected_password_hash: Secret<String>,
+    password_candidate: Secret<String>,
+) -> Result<(), AuthError> {
+    let expected_password_hash = PasswordHash::new(expected_password_hash.expose_secret())
+        .context("Failed to parse has in PHC string format.")?;
+    Argon2::default()
+        .verify_password(
+            password_candidate.expose_secret().as_bytes(),
+            &expected_password_hash,
+        )
+        .context("Invalid password")
+        .map_err(AuthError::InvalidCredentials)
+}
+
+#[tracing::instrument {
+    name ="Retrieve stored credentials",
+    skip(pool)
+    fields(username)
+}]
+async fn get_stored_credentials(
+    username: &str,
+    pool: &SqlitePool,
+) -> Result<Option<(String, Secret<String>)>, anyhow::Error> {
+    let row = sqlx::query!(
+        r#"
+        SELECT user_id, password_hash
+        FROM users
+        WHERE username = $1
+        "#,
+        username,
+    )
+    .fetch_optional(pool)
+    .await
+    .context("Failed to perform a query to retrieve stored credentials.")?
+    .map(|row| (row.user_id, Secret::new(row.password_hash)));
+
+    Ok(row)
+}
+
+#[tracing::instrument {
+    name = "Validate Credentials",
+    skip(credentials, pool)
+}]
+pub async fn validate_credentials(
+    credentials: Credentials,
+    pool: &SqlitePool,
+) -> Result<String, AuthError> {
+    let mut user_id: Option<String> = None;
+    let mut expected_password_hash = Secret::new(
+        "$argon2id$v=19$m=19456,t=2,p=1$QL1KHt3I89l5BGd28zDy+Q$nxQUn9AnAhxcJbQQNa59zlE/JL72RmTre2k/ik5sqC0"
+            .to_string(),
+    );
+
+    if let Some((stored_user_id, stored_password_hash)) =
+        get_stored_credentials(&credentials.username, &pool).await?
+    {
+        user_id = Some(stored_user_id);
+        expected_password_hash = stored_password_hash;
+    }
+
+    let _ = spawn_block_with_tracing(move || {
+        verify_password_hash(expected_password_hash, credentials.password)
+    })
+    .await
+    .context("Failed to spawn blocking task.")??;
+
+    user_id
+        .ok_or_else(|| anyhow::anyhow!("Unknown username."))
+        .map_err(AuthError::InvalidCredentials)
+}
+
+#[tracing::instrument(name = "Change password", skip(password, pool))]
+pub async fn change_password(
+    user_id: &String,
+    password: Secret<String>,
+    pool: &SqlitePool,
+) -> Result<(), anyhow::Error> {
+    let password_hash = spawn_block_with_tracing(move || compute_password_hash(password))
+        .await?
+        .context("Failed to hash password")?;
+    let password_hash_revealed = password_hash.expose_secret();
+    sqlx::query!(
+        r#"
+        UPDATE users
+        SET password_hash = $1
+        WHERE user_id = $2
+        "#,
+        password_hash_revealed,
+        user_id
+    )
+    .execute(pool)
+    .await
+    .context("Failed to change user's password in the database.")?;
+    Ok(())
+}
+
+fn compute_password_hash(password: Secret<String>) -> Result<Secret<String>, anyhow::Error> {
+    let salt = SaltString::generate(&mut OsRng);
+    let password_hash = Argon2::default()
+        .hash_password(password.expose_secret().as_bytes(), &salt)
+        .unwrap()
+        .to_string();
+    Ok(Secret::new(password_hash))
+}
